Correcting For Get_Set Methods
This commit should address getting the EKF origin automatically from the autopilot upon first detection of the heartbeat. The current release of Arducopter will publish both EKF and Home when the Home position is requested. Within this commit there is also handling of message_request however is not currently supported within the 3.6 branch of arducopter.

+++ src/module_vehicle_MAVLINK/controllers/controller_mission.h
@@ -471,9 +471,9 @@ private:
             mace::pose::GeodeticPosition_3D homePosition;
             homePosition.setCoordinateFrame(GeodeticFrameTypes::CF_GLOBAL_AMSL);
             homePosition.setAltitudeReferenceFrame(AltitudeReferenceTypes::REF_ALT_MSL);
-            homePosition.setLatitude(msg.x);
-            homePosition.setLongitude(msg.y);
-            homePosition.setAltitude(msg.z);
+            homePosition.setLatitude(static_cast<double>(msg.x));
+            homePosition.setLongitude(static_cast<double>(msg.y));
+            homePosition.setAltitude(static_cast<double>(msg.z));
 
             std::get<0>(*m_MissionDownloading).setPosition(&homePosition);
             std::get<0>(*m_MissionDownloading).setOriginatingSystem(sender);
